+++ lib/src/model/screen_time_permission_type.dart
@@ -2,6 +2,5 @@ enum ScreenTimePermissionType {
   appUsage,
   accessibilitySettings,
   drawOverlay,
-  queryAllPackages,
   notification,
 }